Refactor: Move analysis and model training pipelines
